+++ test/huawei_in_app_update_test.dart
@@ -8,15 +8,20 @@ void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
 
   setUp(() {
-    channel.setMockMethodCallHandler((MethodCall methodCall) async {
+    final handler = (MethodCall methodCall) async {
       if (methodCall.method == 'checkForUpdate') {
         return {'appId': "1234"};
       }
-    });
+      return null;
+    };
+
+    TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
+        .setMockMethodCallHandler(channel, handler);
   });
 
   tearDown(() {
-    channel.setMockMethodCallHandler(null);
+    TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
+        .setMockMethodCallHandler(channel, null);
   });
 
   test('checkForUpdate', () async {